requirements.txt
pandas>=2.0
requests>=2.28
matplotlib>=3.7
streamlit>=1.31
plotly>=5.18
apscheduler>=3.10
python-kis>=2.1.0
python-dotenv>=1.0
# AI 엔진 의존성
openai>=1.0
google-genai>=1.0
feedparser>=6.0
yfinance>=0.2.30
scikit-learn>=1.3
xgboost>=2.0
joblib>=1.3